Added (dishonorable) autonomous files
Created two files for autonomous mode, one for when the robot starts on the right side of the field, and one for when the robot starts on the left side of the field. Essentially what both the files (should) do is make the robot move in a square so that it pushes samples into the scoring zone. Neither of these files has been tested on the bot as of 10/23/2024.

Issues that may occur with these files include:
1. motors are mapped differently from what I guessed them to be
2. the "sleep" function from util.elapsedtime doesn't work (I remember someone telling me)
3. I did not properly create an opmode
4. I do not know how to program in Java